+++ CHANGELOG.md
@@ -8,11 +8,12 @@ FEATURES:
 * Add option to recursively analyze cookbook's dependencies (defaults to true)
 * Use a file to store configuration info (defaults to *${HOME}/.chef/kitchen_inspector.rb*)
 * Add option to show remarks (*--remarks*), that now are hidden by default
+* Add support for different Repository Managers (but only Gitlab is supported at this stage)
 
 FIXES:
 
 * Handle case in which a given cookbook doesn't exist on both servers
-* Grab information from Chef Server even when it's missing on Gitlab
+* Grab information from Chef Server even when it's missing on the Repository Manager
 
 18 Nov 2013 (1.0.1)
 --